ioapic: add property to disable level interrupt
When it's not possible to intercept eoi to local apic, it's not
possible to emulate level triggered interrupt to re-inject interrupt
when the level is still active.
Prevent trigger mode from setting to be level trigger with
warning with warning once.
Without this guard, qemu can result in unexpected behavior later.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>