i386: Introduce tdx-guest object
Introduce tdx-guest object which implements the interface of
CONFIDENTIAL_GUEST_SUPPORT, and will be used to create a TDX VM by

  qemu -machine ...,confidential-guest-support=tdx0	\
       -object tdx-guset,id=tdx0[,attributes=xxx]

It has only one property 'attributes' so far to configure the attributes
of TDX VM. It's more for developers or advanced users to configure
TD.attributes to whatever value they want.

In the future,

1) separate property of tdx-guest object will be introduced for a
    specific bit. E.g., "debug" for attributes.[bit 0]

2) some bits of the attributes will be tied with flag of other devices.
   E.g., cpu.pmu will be tied with attributes.[bit 63]

Besides, the attributes will be validated to not set any reserved bits
reported by KVM (more specifically TDX module firmware) in the
following patch.

Signed-off-by: Xiaoyao Li <xiaoyao.li@intel.com>