hw/vfio: do the map/unmap at 4K page granularity
Because No guarantee is made to the user that arbitrary unmaps of iova
of size different from those used in the original mapping call will
succeed. Also the actual size of unmap may be different from the
required size.

Use the 4K page as the map/unmap size which is the smallest unit
on Intel platform.

Signed-off-by: Chenyi Qiang <chenyi.qiang@intel.com>