KVM: TDX: Convert page attribute first and then deallocate
When converting memory from private to shared, the current order is,
1. try to deallocate the region by fallocate(PUNCH_HOLE) and 2. tell
KVM to convert the GPA region.  When deallocating, KVM still uses the
GPA region to keep pages.  After that KVM zaps the region on
memory conversion. This causes inefficiency to scan the region twice.

This patch reverses the order. Tell KVM to convert the GPA region to zap
it from GPA mapping to avoid useless scan.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>