i386/tdx: support binding to a MigTD
Pre-migration requires the user TD to be bound to a MigTD. The binding
can be done when
- launching the TD (before TD is finalized), i.e. inital binding; or
- during the TD runtime, i.e. later initial binding or rebinding.

To request binding, the process id of the MigTD is passed to KVM to find
the corresponding MigTD to bind. The binding attributes (defined in TDX
ABI spec) are configurable by users via the "migtd-attr" option. If
user doesn't have one specified, the default attributes will be used.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>